move api into fluid.contrib test=develop
